+++ src/fr/myrddin/clipper/Jni_clipperActivity.java
@@ -13,13 +13,13 @@ public class Jni_clipperActivity extends Activity {
         
         Clipper clipper = new Clipper();
         
-        int[] sp = {325,272,118,272,118,98,325,98};
-        int[] cp = {154,146,274,146,274,231};
+        double[] sp = {325,272,118,272,118,98,325,98};
+        double[] cp = {154,146,274,146,274,231};
         
-        int[][] polygons = clipper.clipPolygon(sp, cp, Clipper.DIFFERENCE);
+        double[][] polygons = clipper.clipPolygon(sp, cp, Clipper.DIFFERENCE);
         
         for (int i = 0; i < polygons.length; i++) {
-            int[] polygon = polygons[i];
+        	double[] polygon = polygons[i];
         	for (int j = 0; j < polygon.length; j++) {
         		Log.e("", "" + polygon[j] );
             }
